refactor(rest): split workspace scaffold modes
