+++ .sampo/changesets/914-split-rest-workspace-scaffold.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split workspace REST resource scaffolding into focused generated-mode, manual-mode, PHP template, and shared type modules while preserving existing add rest-resource behavior.
